change remove button from edit page to details page
